Improved i/o classes in task quadrati
Now input, when used on a file, loads the entire file in memory and then reads it one piece at a time.
Adapted code to C++11 (C++17 is not supported in programming olympiad).
Made all i/o functions always inline, to improve speed (those functions are not large at all).