Improved template to enable (better) debugging.
Added vscode file "launch.json" (it may not work on platforms that do not have gdb debugger).
Added -g to compiler options to enable debugging.
In main.cpp the O3 optimization is enabled only when not debugging (i.e the code is being run on the training website); added -O0 to compiler options to disable every optimization, so that the debugger works better.